requirements.txt
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.2
Flask-Mail==0.10.0
fonttools==4.58.1
fqdn==1.5.1
gcloud==0.18.3
googleapis-common-protos==1.71.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
idna==3.10
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
nest-asyncio==1.6.0
numpy==2.2.6
oauth2client==4.1.3
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prometheus_client==0.22.0
prompt_toolkit==3.0.51
protobuf==6.33.0
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.1
pyparsing==3.2.3
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
python-jwt==4.1.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
rsa==4.9.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.5
